Add details to tab nav
